Add files via upload
Also tried using Word2Vec encoding. But we will use One-hot encoding because it is better than Word2Vec for sarcasm detection in news headlines because it avoids semantic biases inherent in word embeddings and ensures distinct representation for each word, which is crucial for capturing context-specific nature of sarcasm.